+++ src/Mpociot/VatCalculator/Exceptions/VatCheckUnavailableException.php
@@ -4,6 +4,6 @@ namespace Mpociot\VatCalculator\Exceptions;
 
 use Exception;
 
-class VATCheckUnavailableException extends Exception
+class VatCheckUnavailableException extends Exception
 {
 }